Refactor ProfileCards
