+++ src/features/multi-profile/SelectProfileScreen.tsx
@@ -1,32 +1,29 @@
 import { DrawerNavigationProp } from '@react-navigation/drawer';
 import { RouteProp } from '@react-navigation/native';
-import { Card } from 'native-base';
 import React, { Component } from 'react';
-import { Image, SafeAreaView, ScrollView, StyleSheet, TouchableOpacity, View } from 'react-native';
-import key from 'weak-key';
+import { SafeAreaView, ScrollView, StyleSheet, TouchableOpacity, View } from 'react-native';
 
-import { addProfile, NUMBER_OF_PROFILE_AVATARS, tick } from '@assets';
+import { NUMBER_OF_PROFILE_AVATARS } from '@assets';
 import { colors } from '@theme';
-import DaysAgo from '@covid/components/DaysAgo';
 import { Loading, LoadingModal } from '@covid/components/Loading';
 import { Header } from '@covid/components/Screen';
-import { ClippedText, HeaderText, RegularText, SecondaryText } from '@covid/components/Text';
+import { HeaderText, SecondaryText } from '@covid/components/Text';
 import { ApiErrorState, initialErrorState } from '@covid/core/api/ApiServiceErrors';
 import i18n from '@covid/locale/i18n';
-import { AvatarName, getAvatarByName } from '@covid/utils/avatar';
-import { getDaysAgo } from '@covid/utils/datetime';
 import { offlineService, userService } from '@covid/Services';
 import { DrawerToggle } from '@covid/components/DrawerToggle';
+import { ProfileCard } from '@covid/components/ProfileCard';
+import { NewProfileCard } from '@covid/components/NewProfileCard';
 
-import Navigator from '../Navigation';
 import { ScreenParamList } from '../ScreenParamList';
+import Navigator from '../Navigation';
 
 type RenderProps = {
   navigation: DrawerNavigationProp<ScreenParamList, 'SelectProfile'>;
   route: RouteProp<ScreenParamList, 'SelectProfile'>;
 };
 
-type Patient = {
+export type Patient = {
   id: string;
   name?: string;
   avatar_name?: string;
@@ -91,7 +88,7 @@ export default class SelectProfileScreen extends Component<RenderProps, State> {
     try {
       const currentPatient = await userService.getCurrentPatient(patientId);
       this.setState({ isApiError: false });
-      await Navigator.profileSelected(index == 0, currentPatient);
+      await Navigator.profileSelected(index === 0, currentPatient);
     } catch (error) {
       this.setState({
         isApiError: true,
@@ -147,33 +144,17 @@ export default class SelectProfileScreen extends Component<RenderProps, State> {
               {this.state.isLoaded ? (
                 <View style={styles.profileList}>
                   {this.state.patients.map((patient, i) => {
-                    const avatarImage = getAvatarByName((patient.avatar_name ?? 'profile1') as AvatarName);
-                    const hasReportedToday = patient.last_reported_at && getDaysAgo(patient.last_reported_at) === 0;
                     return (
-                      <View style={styles.cardContainer} key={key(patient)}>
+                      <View style={styles.cardContainer} key={patient.id}>
                         <TouchableOpacity onPress={() => this.profileSelected(patient.id, i)}>
-                          <Card style={styles.card}>
-                            <View style={styles.avatarContainer}>
-                              {hasReportedToday && (
-                                <View style={styles.circle}>
-                                  <Image source={tick} style={styles.tick} />
-                                </View>
-                              )}
-                              <Image source={avatarImage} style={styles.avatar} resizeMode="contain" />
-                            </View>
-                            <ClippedText>{patient.name}</ClippedText>
-                            <DaysAgo timeAgo={patient.last_reported_at} />
-                          </Card>
+                          <ProfileCard patient={patient} index={i} />
                         </TouchableOpacity>
                       </View>
                     );
                   })}
 
                   <TouchableOpacity style={styles.cardContainer} key="new" onPress={() => this.gotoCreateProfile()}>
-                    <Card style={styles.card}>
-                      <Image source={addProfile} style={styles.addImage} resizeMode="contain" />
-                      <RegularText>{i18n.t('select-profile-button')}</RegularText>
-                    </Card>
+                    <NewProfileCard />
                   </TouchableOpacity>
                 </View>
               ) : (
@@ -207,50 +188,6 @@ const styles = StyleSheet.create({
     margin: 5,
   },
 
-  avatarContainer: {
-    alignItems: 'center',
-    width: 100,
-    marginBottom: 10,
-  },
-
-  avatar: {
-    height: 100,
-    width: 100,
-  },
-
-  tick: {
-    height: 30,
-    width: 30,
-  },
-
-  circle: {
-    position: 'absolute',
-    alignItems: 'center',
-    justifyContent: 'center',
-    zIndex: 1,
-    top: 0,
-    right: -5,
-    width: 34,
-    height: 34,
-    borderRadius: 17,
-    backgroundColor: 'white',
-  },
-
-  addImage: {
-    width: '100%',
-    height: 100,
-    marginBottom: 10,
-  },
-
-  card: {
-    width: '100%',
-    borderRadius: 16,
-    minHeight: 200,
-    paddingVertical: 20,
-    paddingHorizontal: 12,
-    alignItems: 'center',
-  },
-
   view: {
     backgroundColor: colors.backgroundSecondary,
   },
@@ -261,21 +198,7 @@ const styles = StyleSheet.create({
     justifyContent: 'space-between',
   },
 
-  content: {
-    justifyContent: 'space-between',
-    marginVertical: 32,
-    marginHorizontal: 18,
-  },
-
   rootContainer: {
     padding: 10,
-  },
-
-  shareContainer: {
-    backgroundColor: colors.white,
-    borderRadius: 10,
-    marginHorizontal: 20,
-    paddingHorizontal: 20,
-    paddingVertical: 15,
   },
 });
